+++ src/components/Input.js
@@ -22,6 +22,7 @@ border: 1px solid gray;
 height: 48px;
 font-size: 1rem;
 margin-bottom: 20px;
+outline: none;
 `
 
 Input.propTypes = {
